CV-hw4/requirements.txt
torch>=1.12
torchvision
numpy
opencv-python
matplotlib
pyyaml
cityscapesscripts
tqdm
scikit-learn
pandas
